+++ test/unit-minimal/main.tf
@@ -1,0 +1,37 @@
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# MINIMAL FEATURES UNIT TEST
+# This module tests a minimal set of features.
+# The purpose is to test all defaults for optional arguments and just provide the required arguments.
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+variable "gcp_project" {
+  type        = string
+  description = "(Required) The ID of the project in which the resource belongs."
+}
+
+terraform {
+  required_providers {
+    google = {
+      source  = "hashicorp/google"
+      version = "3.59"
+    }
+  }
+}
+
+provider "google" {
+  project = var.gcp_project
+}
+
+# DO NOT RENAME MODULE NAME
+module "test" {
+  source = "../.."
+
+  # add only required arguments and no optional arguments
+  account_id = "unit-minimal"
+}
+
+# outputs generate non-idempotent terraform plans so we disable them for now unless we need them.
+# output "all" {
+#   description = "All outputs of the module."
+#   value       = module.test
+# }