+++ CHANGELOG.md
@@ -7,11 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.10]
+
+### Changed
+
+- BREAKING: Remove support for Terraform `<= 1.0`. Exclude broken Terraform versions `1.1.0` and `1.1.1`.
+- BREAKING: Update minimum required version of the google provider to `v3.59` to match the minimum requirements of the `terraform-google-service-account-iam` submodule
+
+### Fixed
+
+- Fix attributes in nested `terraform-google-service-account-iam` module to not fail when `module_enabled = false`
+- Wrap `google_service_account` resource in `try` statement in `iam` definition to fix `module_enabled = false`
+- Fix documentation for `projects_access`, `folders_access` and `organization_access`
+
+### Added
+
+- Add unit tests
+
 ## [0.0.9]
 
 ### Fixed
 
-- Fix `module_enaled = false`
+- Fix `module_enabled = false`
 
 ## [0.0.8]
 
@@ -30,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Do not fail unset provider project if project argument is provided
+- Add missing `toset` to `project.roles` and `folder.roles`
 
 ## [0.0.5]
 
@@ -60,12 +78,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Implementation
 
-<!-- markdown-link-check-disable -->
-
-[unreleased]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.9...HEAD
+[unreleased]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.10...HEAD
+[0.0.10]: https://github.com/mineiros-io/terraform-google-service-account/releases/tag/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/mineiros-io/terraform-google-service-account/releases/tag/v0.0.8...v0.0.9
-
-<!-- markdown-link-check-disabled -->
 [0.0.8]: https://github.com/mineiros-io/terraform-google-service-account/releases/tag/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/mineiros-io/terraform-google-service-account/releases/tag/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/mineiros-io/terraform-google-service-account/releases/tag/v0.0.5...v0.0.6